requirements.examples.txt
# requirements to run example.py and fastapi-example.py
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-semantic-conventions>=0.40b0
aiohttp>=3.9.0

fastapi
dotenv
openai>=1.0.0
ddgs>=5.0.0

